requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.49
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.13.3
orjson==3.10.1
packaging==23.2
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
PyPDF2==3.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
yarl==1.9.4
